+++ src/services/workspaceService.ts
@@ -2,7 +2,6 @@
  * Service for detecting multiple projects in workspace
  */
 
-import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';